# social_graph_service/src/config/swaggerDoc.yaml
openapi: 3.0.1
info:
  title: moody Social Graph Service
  description: This is the REST API documentation for the moody social_graph_service
  version: 1.0.0
servers:
  - url: /api
...
paths:
  /profiles:
    post:
      tags:
        - profile
      summary: Create a new profile or update an existing one
      description: If the user (provided via the header) does not have a profile yet, a new profile will be created. Else the profile will be updated.
      requestBody:
        description: Profile object to be created or updated
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Profile'
        required: true
      responses:
        201:
          description: Profile created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 201
                  profile:
                    $ref: '#/components/schemas/Profile'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 401
                  msg:
                    type: string
                    example: Unauthorized - Token is not valid
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 500
                  msg:
                    type: string
                    example: Internal Server Error
      security:
        - authentication: []
    get:
      tags:
        - profile
      summary: Get all profiles
      description: Fetching all exisiting profiles.
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 200
                  profiles:
                    type: array
                    items:
                      $ref: '#/components/schemas/Profile'
        404:
          description: No profiles found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 404
                  msg:
                    type: string
                    example: No profiles found
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 500
                  msg:
                    type: string
                    example: Internal Server Error
    delete:
      tags:
        - profile
      summary: Delete all registered profiles
      description: all existing profiles are deleted (for testing purpose only)
      responses:
        200:
          description: All profiles removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 200
                  msg:
                    type: string
                    example: All profiles removed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 500
                  msg:
                    type: string
                    example: Internal Server Error
...
  /follow/{username}:
    post:
      tags:
        - followerRelation
      summary: Create a new follow relationship between two profiles
      description: a relationship will be created betweent two existing moody profiles.
      parameters:
        - in: path
          name: username
          schema:
            type: string
            example: Birgit
          required: true
          description: the username of a profile
      responses:
        201:
          description: Follow relationship created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 201
                  relationship:
                    type: string 
                    example: follows
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 401
                  msg:
                    type: string
                    example: Unauthorized - Token is not valid
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 500
                  msg:
                    type: string
                    example: Internal Server Error
      security:
        - authentication: []
    get:
      tags:
        - followerRelation
      summary: Check follow relationship
      description: Check if a specific follow relationship between the current user and an other user exists
      parameters:
        - in: path
          name: username
          schema:
            type: string
            example: Birgit
          required: true
          description: the username of a profile
      responses:
        200:
          description: true if relationship exists and false if not
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number 
                    example: 200
                  following:
                    type: boolean  
                    example: true
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 401
                  msg:
                    type: string
                    example: Unauthorized - Token is not valid
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 500
                  msg:
                    type: string
                    example: Internal Server Error
      security:
        - authentication: []
    delete:
      tags:
        - followerRelation
      summary: Delete an existing follow relationship between two profiles
      description: Delete an existing follow relationship between two profiles
      parameters:
        - in: path
          name: username
          schema:
            type: string
            example: Birgit
          required: true
          description: the username of a profile
      responses:
        201:
          description: Follow relationship deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 201
                  relationship:
                    type: string 
                    example: unfollows
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 401
                  msg:
                    type: string
                    example: Unauthorized - Token is not valid
        404:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 404
                  msg:
                    type: string
                    example: Relationship does not exist
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 500
                  msg:
                    type: string
                    example: Internal Server Error
      security:
        - authentication: []
...
components:
  schemas:
    Profile:
      required:
        - birthday
        - hometown
        - gender
      type: object
      properties:
        birthday:
          type: string
          format: date
        hometown:
          type: string
          example: Linz
        gender:
          type: string
          example: female
          enum:
            - male
            - female
            - transgender
        privacy:
          type: boolean
          example: false
          default: false
        notifications:
          type: boolean
          example: false
          default: false
        username:
          type: string
          example: Hugo
...
  securitySchemes:
    authentication:
      type: apiKey
      name: Authorization
      in: header